% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisaClust.R
\name{lisaClust}
\alias{lisaClust}
\title{Use k-means clustering to cluster local indicators of spatial association. For other clustering use lisa.}
\usage{
lisaClust(
  cells,
  k = 2,
  r = NULL,
  imageID = "imageID",
  cellType = "cellType",
  spatialCoords = c("x", "y"),
  regionName = "region",
  cores = 1,
  window = "convex",
  window.length = NULL,
  whichParallel = "imimageID",
  sigma = NULL,
  lisaFunc = "K",
  minLambda = 0.05,
  BPPARAM = NULL,
  Rs = r
)
}
\arguments{
\item{cells}{A SingleCellExperiment, SpatialExperiment or data frame that contains at least the
variables x and y, giving the  coordinates of each cell, imageID and cellType.}

\item{k}{The number of regions to cluster.}

\item{r}{A vector of the radii that the measures of association should be calculated.}

\item{imageID}{The column which contains image identifiers.}

\item{cellType}{The column which contains the cell types.}

\item{spatialCoords}{The columns which contain the x and y spatial coordinates.}

\item{regionName}{The output column for the lisaClust regions.}

\item{cores}{Number of cores to use for parallel processing, or a BiocParallel 
MulticoreParam or SerialParam object.}

\item{window}{Should the window around the regions be 'square', 'convex' or 'concave'.}

\item{window.length}{A tuning parameter for controlling the level of concavity
when estimating concave windows.}

\item{whichParallel}{Should the function use parallization on the imageID or
the cellType.}

\item{sigma}{A numeric variable used for scaling when filting inhomogeneous L-curves.}

\item{lisaFunc}{Either "K" or "L" curve.}

\item{minLambda}{Minimum value for density for scaling when fitting inhomogeneous L-curves.}

\item{BPPARAM}{\{DEPRECATED\} A BiocParalell MulticoreParam or SerialParam object.}

\item{Rs}{A vector of the radii that the measures of association should be calculated.}
}
\value{
A matrix of LISA curves
}
\description{
Use k-means clustering to cluster local indicators of spatial association. For other clustering use lisa.
}
\examples{
## Generate toy data: two images, two cell types that sit in separate bands
set.seed(51773)
x <- round(c(
  runif(200), runif(200) + 1, runif(200) + 2, runif(200) + 3,
  runif(200) + 3, runif(200) + 2, runif(200) + 1, runif(200)
), 4) * 100
y <- round(c(
  runif(200), runif(200) + 1, runif(200) + 2, runif(200) + 3,
  runif(200), runif(200) + 1, runif(200) + 2, runif(200) + 3
), 4) * 100
cellType <- factor(paste("c", rep(rep(c(1:2), rep(200, 2)), 4), sep = ""))
imageID <- rep(c("s1", "s2"), c(800, 800))
cells <- data.frame(x, y, cellType, imageID)

# Cluster the cells into two regions
cells <- lisaClust(cells, k = 2)
table(cells$region, cells$cellType)

}
