% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisaClust.R
\name{lisaClust}
\alias{lisaClust}
\title{Use k-means clustering to cluster local indicators of spatial association. For other clustering use lisa.}
\usage{
lisaClust(
  cells,
  k = 2,
  Rs = NULL,
  spatialCoords = c("x", "y"),
  cellType = "cellType",
  imageID = "imageID",
  regionName = "region",
  BPPARAM = BiocParallel::SerialParam(),
  window = "convex",
  window.length = NULL,
  whichParallel = "imageID",
  sigma = NULL,
  lisaFunc = "K",
  minLambda = 0.05,
  fast = TRUE
)
}
\arguments{
\item{cells}{A SegmentedCells, SingleCellExperiment, SpatialExperiment or data frame that contains at least the 
variables x and y, giving the  coordinates of each cell, imageID and cellType.}

\item{k}{The number of regions to cluster.}

\item{Rs}{A vector of the radii that the measures of association should be calculated.}

\item{spatialCoords}{The columns which contain the x and y spatial coordinates.}

\item{cellType}{The column which contains the cell types.}

\item{imageID}{The column which contains image identifiers.}

\item{regionName}{The output column for the lisaClust regions.}

\item{BPPARAM}{A BiocParallelParam object.}

\item{window}{Should the window around the regions be 'square', 'convex' or 'concave'.}

\item{window.length}{A tuning parameter for controlling the level of concavity 
when estimating concave windows.}

\item{whichParallel}{Should the function use parallization on the imageID or 
the cellType.}

\item{sigma}{A numeric variable used for scaling when filting inhomogeneous L-curves.}

\item{lisaFunc}{Either "K" or "L" curve.}

\item{minLambda}{Minimum value for density for scaling when fitting inhomogeneous L-curves.}

\item{fast}{A logical describing whether to use a fast approximation of the 
inhomogeneous local L-curves.}
}
\value{
A matrix of LISA curves
}
\description{
Use k-means clustering to cluster local indicators of spatial association. For other clustering use lisa.
}
\examples{
library(spicyR)
# Read in data as a SegmentedCells objects
isletFile <- system.file("extdata","isletCells.txt.gz", package = "spicyR")
cells <- read.table(isletFile, header=TRUE)
cellExp <- SegmentedCells(cells, cellProfiler = TRUE)

# Cluster cell types
markers <- cellMarks(cellExp)
kM <- kmeans(markers,8)
cellType(cellExp) <- paste('cluster',kM$cluster, sep = '')

# Generate LISA
cellExp <- lisaClust(cellExp, k = 2)

}
